refactor: gui 에서 사용하지 않는 type 정리
