+++ apps/gui/src/renderer/hooks/queries/use-mcp.ts
@@ -1,6 +1,6 @@
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
-import { Services } from '../../bootstrap';
 import type { McpConfig } from '@agentos/core';
+import { ServiceContainer } from '../../../shared/ipc/service-container';
 
 // Query Keys
 const QUERY_KEYS = {
@@ -11,7 +11,7 @@ const QUERY_KEYS = {
 
 // MCP 설정 목록 조회
 export const useMcpConfigs = () => {
-  const mcpService = Services.getMcp();
+  const mcpService = ServiceContainer.getOrThrow('mcp');
 
   return useQuery({
     queryKey: QUERY_KEYS.mcpConfigs,
@@ -22,7 +22,7 @@ export const useMcpConfigs = () => {
 
 // MCP 상태 조회 (특정 클라이언트)
 export const useMcpStatus = (clientName: string) => {
-  const mcpService = Services.getMcp();
+  const mcpService = ServiceContainer.getOrThrow('mcp');
 
   return useQuery({
     queryKey: [...QUERY_KEYS.mcpStatus, clientName],
@@ -36,7 +36,7 @@ export const useMcpStatus = (clientName: string) => {
 // MCP 연결 뮤테이션
 export const useConnectMcp = () => {
   const queryClient = useQueryClient();
-  const mcpService = Services.getMcp();
+  const mcpService = ServiceContainer.getOrThrow('mcp');
 
   return useMutation({
     mutationFn: (config: McpConfig) => mcpService.connectMcp(config),
@@ -50,7 +50,7 @@ export const useConnectMcp = () => {
 // MCP 연결 해제 뮤테이션
 export const useDisconnectMcp = () => {
   const queryClient = useQueryClient();
-  const mcpService = Services.getMcp();
+  const mcpService = ServiceContainer.getOrThrow('mcp');
 
   return useMutation({
     mutationFn: (configId: string) => mcpService.disconnectMcp(configId),
@@ -64,7 +64,7 @@ export const useDisconnectMcp = () => {
 // MCP 설정 저장 뮤테이션 (connect로 대체)
 export const useSaveMcpConfig = () => {
   const queryClient = useQueryClient();
-  const mcpService = Services.getMcp();
+  const mcpService = ServiceContainer.getOrThrow('mcp');
 
   return useMutation({
     mutationFn: (config: McpConfig) => mcpService.connectMcp(config),
@@ -77,7 +77,7 @@ export const useSaveMcpConfig = () => {
 // MCP 설정 삭제 뮤테이션 (disconnect로 대체)
 export const useDeleteMcpConfig = () => {
   const queryClient = useQueryClient();
-  const mcpService = Services.getMcp();
+  const mcpService = ServiceContainer.getOrThrow('mcp');
 
   return useMutation({
     mutationFn: (configId: string) => mcpService.disconnectMcp(configId),
