+++ apps/gui/src/renderer/hooks/useAppData.ts
@@ -1,8 +1,7 @@
 import { useState, useEffect } from 'react';
-import type { UseAppDataReturn } from '../types/core-types';
-import { PresetServiceInterface, McpServiceInterface } from '../types/core-types';
+import type { UseAppDataReturn } from '../stores/store-types';
 import { Preset, McpConfig, ReadonlyAgentMetadata, ReadonlyPreset } from '@agentos/core';
-import { ServiceContainer } from '../services/service-container';
+import { ServiceContainer } from '../../shared/ipc/service-container';
 
 /**
  * App data management hook
@@ -24,7 +23,7 @@ export function useAppData(): UseAppDataReturn {
         console.log('🔄 Loading presets from PresetService...');
 
         if (ServiceContainer.has('preset')) {
-          const presetService = ServiceContainer.get<PresetServiceInterface>('preset');
+          const presetService = ServiceContainer.getOrThrow('preset');
           console.log('📦 PresetService found, calling getAllPresets()...');
 
           const corePresets = await presetService.getAllPresets();
@@ -99,7 +98,7 @@ export function useAppData(): UseAppDataReturn {
       console.log('🔄 Creating new preset:', newPresetData);
 
       if (ServiceContainer.has('preset')) {
-        const presetService = ServiceContainer.get<PresetServiceInterface>('preset');
+        const presetService = ServiceContainer.getOrThrow('preset');
 
         const presetToCreate: Preset = {
           id: `preset-${Date.now()}`,
@@ -126,7 +125,7 @@ export function useAppData(): UseAppDataReturn {
         };
 
         console.log('📤 Sending preset to service:', presetToCreate);
-        const result = await presetService.createPreset(presetToCreate as any);
+        const result = await presetService.createPreset(presetToCreate);
         console.log('📥 Service create result:', result);
 
         setPresets((prev) => [...prev, result]);
@@ -145,7 +144,7 @@ export function useAppData(): UseAppDataReturn {
   const handleCreateMCPTool = async (mcpConfig: McpConfig): Promise<unknown> => {
     try {
       if (ServiceContainer.has('mcp')) {
-        const mcpService = ServiceContainer.get<McpServiceInterface>('mcp');
+        const mcpService = ServiceContainer.getOrThrow('mcp');
         await mcpService.connectMcp(mcpConfig);
         return mcpConfig;
       }
@@ -219,7 +218,7 @@ export function useAppData(): UseAppDataReturn {
       console.log('🔄 Updating preset:', updatedPreset);
 
       if (ServiceContainer.has('preset')) {
-        const presetService = ServiceContainer.get<PresetServiceInterface>('preset');
+        const presetService = ServiceContainer.getOrThrow('preset');
 
         // DesignPreset을 Core Preset으로 변환하여 업데이트
         const corePreset: Preset = {
@@ -251,7 +250,7 @@ export function useAppData(): UseAppDataReturn {
       console.log('🔄 Deleting preset:', presetId);
 
       if (ServiceContainer.has('preset')) {
-        const presetService = ServiceContainer.get<PresetServiceInterface>('preset');
+        const presetService = ServiceContainer.getOrThrow('preset');
 
         console.log('📤 Sending delete request to service for:', presetId);
         const result = await presetService.deletePreset(presetId);
