+++ apps/gui/src/renderer/hooks/useAppNavigation.ts
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import type { AppSection, UseAppNavigationReturn } from '../types/core-types';
+import type { AppSection, UseAppNavigationReturn } from '../stores/store-types';
 import { ReadonlyPreset } from '@agentos/core';
 
 /**
