--- apps/gui/src/renderer/services/index.ts
@@ -1,14 +1,0 @@
-// Export all service classes (DI 기반)
-export { BridgeService } from './bridge.service';
-export { McpService } from './mcp-service';
-export { PresetService } from './preset-service';
-
-// ServiceContainer
-export { ServiceContainer } from './service-container';
-
-// IPC Channel exports
-export * from './ipc';
-
-// Bootstrap exports
-export { bootstrap, Services, isBootstrapped, shutdown } from '../bootstrap';
-export type { BootstrapResult } from '../bootstrap';
