+++ apps/gui/src/shared/ipc/impl/mock-ipc-channel.ts
@@ -11,7 +11,7 @@ import type {
   McpUsageStats,
   Preset,
 } from '@agentos/core';
-import type { IpcChannel } from '../../../shared/types/ipc-channel';
+import type { IpcChannel } from '../../types/ipc-channel';
 import type {
   CursorPagination,
   CursorPaginationResult,
@@ -22,7 +22,7 @@ import type {
   ResourceListResponse,
   ResourceResponse,
   ToolExecutionResponse,
-} from '../../types/core-types';
+} from '../../types/ipc-channel';
 import { LlmManifest, UserMessage, Message } from 'llm-bridge-spec';
 import type {
   ClearUsageLogsResponse,
@@ -30,7 +30,7 @@ import type {
   McpUsageUpdateEvent,
   SetUsageTrackingResponse,
   UsageLogQueryOptions,
-} from '../../../shared/types/mcp-usage-types';
+} from '../../types/mcp-usage-types';
 
 export class MockIpcChannel implements IpcChannel {
   private bridges = new Map<string, LlmManifest>();
